Increase compiler warnings flags

+++ hw/bsp/board.c
@@ -135,13 +135,13 @@ TU_ATTR_USED int sys_read (int fhdl, char *buf, size_t count)
 TU_ATTR_USED int sys_write (int fhdl, const void *buf, size_t count)
 {
   (void) fhdl;
-  return board_uart_write(buf, count);
+  return board_uart_write(buf, (int) count);
 }
 
 TU_ATTR_USED int sys_read (int fhdl, char *buf, size_t count)
 {
   (void) fhdl;
-  return board_uart_read((uint8_t*) buf, count);
+  return board_uart_read((uint8_t*) buf, (int) count);
 }
 
 #endif
